Files added -- rerun configure.
Eager to test out our support for existential types, I have added a
new library module, iter.  Here is the entirety of iter.h:

namespace Iter {
  struct Iter<`a::A,`eff::E> {
    <`env>
    `env env;
    // return false when no more elts.
    bool (@next)<`r::R>(`env env, `a @`r dest; `eff + {`r});
  };
  typedef struct Iter<`a,`eff> iter_t<`a,`eff>;
  bool next(iter_t<`a::A>,`a::A @);
}

This is not necessarily the best design -- you could imagine an
is_empty method, next throwing an exception on "no more elements" and
next returning an `a.  But then `a would need kind B, or next would
need to be passed a region handle and return `a @`r.  This way the
result (*dest) can be stack-allocated.

set and dict provide iterators:

Iter::iter_t<`a,{`r1,`r2}> Set::make_iter(region_t<`r1>, set_t<`a,`r2>);
Iter::iter_t<$(`a,`b),{`r1,`r2}> Dict::make_iter(region_t<`r1>,dict_t<`a,`b,`r2>);

The region passed to the iterator-makers is for the iterators' private
storage, which is why it shows up in the effect.  Note that explicit
effects seem necessary in the result types.  Also notice
Dict::make_iter exploits the kind A on the `a in the definition of
struct Iter.

Notice that the iterator maker makes no commitment on where the
results of calls to the next "method" are stored -- hence the
higher-order polymorphism in the definition of struct Iter.

With this, I was able to eliminate the abstraction-breaking
Set::elements function.  You can use an iterator instead, although it
isn't quite as convenient.  Here for example is the modified
prefix_of_member function in cf_flowinfo.cyc:

static bool prefix_of_member(region_t<`r> r, place_t<`r2,`r3> place,
			     Set::set_t<place_t<`r2,`r3>> set) {
  let place2 = place;
  let iter = Set::make_iter(r,set);
  while(Iter::next(iter,&place2)) { // was a for loop over Set::elements(set)
     ...rest of function essentially unchanged...

Unfortunately we had to give an initializer to place2.  An
"initializes" qualifier isn't quite enough because Iter::next
initializes place2 only when it returns true.  This is a common C
idiom I would very much like to support.  Initializing with place
required the `r2 and `r3 equalities in the parameters -- they are not
otherwise necessary.

The storage used by a set iterator is O(1).  The storage used by a
dict iterator is O(log n) for a dict with n elements (its
implementation is pretty clever if I do say so myself).  The Dict
iterator is used somewhat gratuitously in the treeshaking in tc.cyc.

To implement these iterators, I implemented partial type
instantiation.  Let's walk through the set iterator to see why:
bool iter_f<`a,`r1,`r2,`r3>(list_t<`a,`r1> @`r2 elts_left, `a @`r3 dest;
			    {`r1,`r2,`r3}) {
  if(!*elts_left)
    return false;
  *dest = (*elts_left)->hd;
  *elts_left = (*elts_left)->tl;
  return true;
}
Iter::iter_t<`a,{`r1,`r2}> make_iter(region_t<`r1> rgn, set_t<`a,`r2> s) {
  return Iter::Iter(rnew(rgn) s->nodes, iter_f@<`a,`r2,`r1><>);
}

To get iter_f to unify with Iter::Iter.next, we need to instantiate
iter_f's `a, `r1, and `r2, but not `r3.  It was easy enough to let the
type-checker accept iter_f@<`a,`r2,`r1>, but it then went and finished
the instantiation implicitly!  So I relaxed the parser to allow
multple instantiation/no-instantiation things.  But probably the
default should be that iter_f@<`a,`r`,`r1> means
iter_f@<`a,`r2,`r1><>.

As you can see, it took some pretty hefty type acrobatics to push all
of this through.  We'll see how much region bounds a la the PLDI
submission help matters -- my guess is not much.  In particular, it's
frustrating the iterator-providers need to be so clever.

If you'd like to play with this stuff, here are some suggestions:
* replace some uses of Set::iter[_c] or Dict::iter[_c]
* implement Hashtable::make_iter
* make it so fewer explicit type information is necessary

Other (related) changes:
* Had to add some more "casts to shut up gcc warnings" with respect to
initializing struct fields of struct types that had existential
types.  There are no changes to non-existential structs.

* We were never using Dict::choose, which isn't surprising since it
  wasn't implemented.  We implemented Dict::rchoose -- dict.h has been
  updated to reflect this.  Link-checker anyone?

* Changed cyctest.cyc because Set::elements no longer exists.  Instead
  we use an iterator to implement set_elements (outside the module, if
  you will).  If you look at these tests, though, they are bogus
  because they assume the order of Set::elements -- I put in a reverse
  so the tests pass, but there's still no reason to think they should.

Question (for Greg, I presume):
* Why are the ::A necessary in the prototype of Iter::next ?


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1847 581db07f-6e62-4d40-9733-d432c4eebb07