1. added support for abstract typedefs.  You simply use "_" in place of the definition.  For instance:
   typedef _ foo_t;  // foo_t is now an abstract type -- default is BoxKind
   typedef _::A region_t<`r::R>; // region_t is an abstract type constructor
                                 // with argument of kind R and returns an
                                 // abstract type.

The problem is that, currently, you can't define a typedef more than
once.  So this is good for interfaces being exported to clients, but
not to be #included into the file that defines it.  The problem is
a parsing one (surprise) because you can't have the same identifier
showing up as the defined identifier in two typedefs.

2. check that main has a reasonable type -- it must at least return
an int, and if it has one argument, then it must be an int, and if
it has 2 arguments, the second one has to be char?? or char?`r1?`r2
or const char?`r1?`r2 etc.  This code should really go in the link
checker when we finally get that integrated but it avoids the problem
that one of the reviewers ran into when he tried writing "char **"
for argv instead of "char ??".

3. added support for "resetable" regions.  You can declare a dynamic
region using "region [resetable] r1 { ... }" and within the scope of
the region, call reset_region(r1).  This allows you to free the pages
associated with the region, and then continue to allocate new stuff
within the region.  The control-flow analysis ensures that you never
dereference the "old" pointers.

This allows us to code up the "interpreter" example with no space
leaks.

extern roots_t<`r> init_roots(region_t<`r>);
extern roots_t<`r2> copy_roots(region_t<`r2>, roots_t<`r1>);
extern roots_t<`r> compute(region_t<`r>, roots_t<`r>);

void interp() {
  // r1 and r2 correspond to "from" and "to" space
  region [resetable] r1 {
  region [resetable] r2 {
    // create initial roots in r1
    roots_t<`r1> r1_roots = init_roots(r1);
    roots_t<`r2> r2_roots;
    // interpreter loop
    while (true) {
      // compute in r1 for a while until gc is required
      r1_roots = compute(r1,r1_roots);
      // copy live data from r1 into r2
      r2_roots = copy_roots(r2,r1_roots);
      // free data in r1
      reset_region(r1);
      // r1_roots is now uninitialized as far as analysis goes
      // compute in r2 for a while until gc is required
      r2_roots = compute(r2,r2_roots);
      // copy live data from r2 into r1
      r1_roots = copy_roots(r1,r2_roots);
      // free data in r2
      reset_region(r2);
      // r2_roots is uninitialized as far as analysis goes
    }
  }}
}

A few technical details on how this works:  When reseting a region,
we simply make all of the variables that might point into that region
uninitialized.  More specifically, any variable (or root) that has
in its type the region being reset is reset to uninitialized in the
flow analysis.  This ensures that you can't use, for instance,
r1_roots after resetting r1 until you assign something else to r1
(e.g., via copy_roots.)

To further ensure soundness, we do not allow you to cast a pointer
t@`r1 to t@`r2 when r1 is a resetable region even if `r2 is declared
within the scope of `r1.  The reason, of course, is that you could
then hide a pointer as a `r2 pointer and then dereference it after
resetting `r1.  More specifically, a resetable region is never an
ancestor in the "outlives" relationship maintained by the analysis.
[This isn't quite true -- if we pass r1 to a procedure, and it
declares a new region r2, then r1 will be seen as an ancestor of r2.
What saves us here is that within the procedure, you can't reset r1.]

In turn, this means that (once we have outlives constraints on
existentials as in the paper), you won't be able to open a package
with a bound (such as `r2) when the hidden region is `r1.

Finally, we only allow reset within the scope of the region
definition.  This ensures that there is no aliasing among
the region in question.

We can pretty easily extend this to an interprocedural setup
by having a "resets(`r)" effect.  This will require that `r
not be aliased with another region, and that `r be resetable
and so forth.  In this fashion, we can have early deallocation
as we were planning earlier, and the reset functionality.

Note that early deallocation is not sufficient to code the
interpreter above -- you still end up piling up region handles
(either on the stack or some other region).  Resetting gives
you the "tail-call" you need for regions.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1856 581db07f-6e62-4d40-9733-d432c4eebb07