I'm not sure why so many genfiles changed.  Hopefully this is not a bad sign.  In any case, I made the following changes:
1) Added function unsigned int Core::arr_prevsize(`a::A ?) that returns
the length from the current location to the beginning of the buffer of
the given ? value.  Added to the array test in cyctest to test this.

2) Started to construct code that will marshal the interface environment
used for interface checking.  I got as far as creating marshalling
routines for dict_t, and for setting up some marshalling routines for
interfaces.  Didn't finish though, so in effect the interface checker
isn't any further along.  Will play with it as time goes on.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1878 581db07f-6e62-4d40-9733-d432c4eebb07