Fixed a bug in cycflex: The result of tmpnam(NULL) is a pointer to a static buffer in memory, so after doing
x = tmpnam(NULL)
write to x...
y = tmpnam(NULL)
write to y...
do something involving both x and y...

x and y are both the same filename.  In the case of flex, this meant we
were doing flex -Stmp2 -otmp2 infile.l rather than flex -Stmp1 -otmp2
infile.l.  The output of flex in such circumstances is (a) incorrect and
(b) infinite.

I'm not sure why this worked before, but I conjecture that the
buildlib-ified libraries implement tmpnam "correctly" (ie less safely)
than the original Cyclone-wrapped one did.

The right thing to do is either use mkstemp (which we don't seem to
support yet, correct me if I'm wrong) or to guard the uses of tmpnam
with strdup (which is what I did.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2024 581db07f-6e62-4d40-9733-d432c4eebb07