Fixed the bug in the occurs-check that Dan found.
Made a few changes to the front-end to allow more C code to
parse for buildlib.  The changes are:

* allow empty list of members in struct/union declarations.

* allow parser to treat Cyclone-added keywords as identifiers. Note
that I modified buildlib so that it uses the parser with this feature
enabled.

* catch some errors in the parser so it's more likely we recover.

The last change involved adding some error cases to the parser
and catching exceptions that were thrown during certain downcasts
(using the $i notation for Cycbison.)  This should have the side
effect that we have fewer uncaught exceptions when attempting
to compile buggy code and should allow (in some situations) the
parser to recover enough to skip over certain declarations.

Anyway, the insertion of the error productions produced a 2 new
shift/reduce conflicts so right now, we have a total of 5 of these
buggers:  1 from if-then-else, 2 from Trevor's recent changes,
and 2 from my changes here.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2109 581db07f-6e62-4d40-9733-d432c4eebb07