Added a -Wall (warn all) flag which now warns about both array bounds checks and null checks that are inserted by the compiler.
Fixed a bug in the control-flow analysis (thanks Dan) that did
not treat initialization correctly for variables with @ type.

For fun, got rid of all of the null checks inserted by the compiler
in list.cyc by introducing the appropriate temporaries or using
@ types more carefully.  There's one null check left but it's
pretty hard to get rid of since it involves reasoning about the
length of a list and some side-effects within a comprehension.

Got rid of a few other null checks in the other changed files.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2193 581db07f-6e62-4d40-9733-d432c4eebb07