* More flexibility about where attributes can appear.  In particular, now you can have:  int (__attribute__((noreturn)) *f)(int).  Unfortunately, This pushes us up to 11 shift/reduce conflicts.
* Finally fixed the problem that qualifiers are lost within
typedefs.  Now the typedef declaration carries the information with it.
In addition, tqual_t is changed to have print_const and real_const fields
in lieu of the q_const field.  The real_const field determines whether
or not the object is read-only and is true if either print_const is
true or else the object's type is a typedef whose real_const is true.
The print_const is only set by the user writing "const" explicitly.
I had to modify everywhere we construct a tqual to look at the element
type and extract the "real" qualifier from a combination of the
print_const and any typedefs.  As a result, you can write something
like:

typedef const int cint_t;
cint_t x;
struct Foo {
  cint_t y;
};
struct Foo z;

and now x and z.y are read-only.  The reason for the distinction
between "real" and "print" const is that we can't just translate
the declaration of x above to:

const cint_t x;

because GCC complains about the extra const.  But we need to record
somewhere that x is read-only so that the type-checker can enforce
the constraint.  Hence, each tqual has what to print and what the
real qualifier is.  Note that I don't have this distinction for volatile
and restrict simply because the type-checker doesn't need to enforce
anything for these qualifiers.

I'm about half-way done getting attributes in typedef's right and
will check that in soon.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2220 581db07f-6e62-4d40-9733-d432c4eebb07