* Got rid of codegen, splice, cut, fill, forarray, and switch "C" since we either don't use them or they are no longer needed.  The goal is to cut down on keywords and simplify the abstract syntax.
* Added a new construct:

extern "C include" { <decls> } export { x1,...,xn;}.

This construct allows you to put arbitrary C code in for <decls>.  For
instance, one can write:

extern "C include" {
  struct Foo { void *x; };

  typedef struct Foo *foo_t;

  int *bar(foo_t f) {
    return (int *)(f->x);
  }

  int foo(foo_t f) {
    return *bar(f);
  }
} export {
  foo;
}

What happens is that we type-check all of the type definitions, as
well as the types for variable declarations and the prototypes for
function declarations.  However, we do not check initializers for
variables nor function bodies.  We enter all type definitions into the
environment, but we only enter a variable definition or function
definition if it is explicitly exported.  So, any code following the
above declaration can use struct Foo, foo_t and foo(), but can't use
bar.

The translation to C preserves all of the definitions within the
braces and spits it out untranslated.  So, this is a convenient way to
include C code within Cyclone code.  For instance, one can write:

extern "C include" {
#include <jpeglib.h>
} export {
  jpeg_CreateCompress, jpeg_CreateDecompress,
  jpeg_destroy_compress, jpeg_destroy_decompress;
}

to get a quick and dirty interface on the jpeg library.  Of course,
you're better off not exporting these functions with the C prototypes,
but rather re-declaring them with appropriate Cyclone prototypes.

Right now, the limiting factor for including arbitrary C code is the
parser.  In particular, we need to (a) deal with the typedef problem,
and (b) add asm and other GCC features to the abstract syntax in order
to allow #include'ing more C header files.  However, I was happy to
see that I could include a large number of header files from Cygwin,
such as crypt.h, jpeglib.h, mapi.h, pcre.h, pthread.h, tcl.h, and
windows.h.  This is a really convenient way to suck all of the type
definitions for these libraries into Cyclone.

It's worth noting that if you happen to redeclare a function or
variable using Cyclone types, then you'll get the benefit of C
checking that the definition is compatiable (at the C level)
with the things that you put in the extern "C include" definitions.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2242 581db07f-6e62-4d40-9733-d432c4eebb07