Added new syntactic form for use with unique pointers:
alias e = <`r>x s

That is: assuming expression e is a unique pointer, bind it to the
variable x and given the pointer region annotation `r for the
duration of statement s.  Within s, the location e is considered
consumed (i.e. not allowed to be accessed), but it is then
unconsumed following s.  At the moment, e is restricted to being
a non-path (i.e. a variable), but hopefully this restriction will
be relaxed soon.

The main drawback with this construct at the moment is that it
is not a "deep" coercion.  For example, if I had

  struct List<int,`U> *`U x = ...
  alias x = <`r>y {
    ...
  }

Then within the alias statement, y would have type struct List<int,`U> *`r
rather than (the preferred) struct List<int,`r> *`r.  Doing the
coercion in this case is unsound because we might create aliases
when using y, but it could be made sound if the fields that we
are deeply coercing are const.  I'll look to at least add this
much later.

Also fixed some bugs:
1) Wasn't dealing with assignments to unique paths in the flow analysis.
2) Atomic_effect_outlives wasn't compressing
   a type, so the outlives result was incorrect; my mistake
   from before.
3) is_unique_path wasn't properly dealing with cast expressions
   and was erroneously considering NULL not to be a unique path.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2289 581db07f-6e62-4d40-9733-d432c4eebb07