A few changes to get things to compile on OS X:
Some code in tools/cyclex/compact.cyc was triggering a bug
in -elim-statement-expressions (it was an explicit statement
expression inside a comprehension, leading to gotos inside
a statement expression in the output of toc).  I changed
the code of compact.cyc to get around the bug; the bug
remains to be fixed. (!)

Changed how _INLINE_FUNCTIONS gets communicated to
cyc_include.h; this means that cyc_include.h and therefore
all genfiles have changed.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2298 581db07f-6e62-4d40-9733-d432c4eebb07