Got rid of the '#include <setjmp.h>' in cyc_include.h. This was polluting the namespace in such a way that we could not support signal on cygwin.  Now signal is supported on cygwin.
Added a flag, -setjmp, to buildlib.  This extracts the
minimal declarations needed to use setjmp on a system,
namely, a declaration for setjmp and jmp_buf, plus whatever
those require.  (On Linux this is actually nontrivial.)
We use this to produce a file, cyc_setjmp.h, that is
prepended to every C file we generate from Cyclone files.
(We need the declaration of setjmp in generated files
because we use it for implementing exceptions.)

Every genfile has changed.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2304 581db07f-6e62-4d40-9733-d432c4eebb07