Fixed up alias construct to do "deep" aliasing.  Here's how it works.  The construct has syntax
alias <`r>x = e in s

Here `r and x are being defined, where x is bound in s and `r is
bound in e and s. The region variable `r is non-linear.  Furthermore,
if e is a unique pointer, it is consumed for the duration of s,
but then unconsumed once s is complete.

To use this construct to "open" a unique pointer, you include a cast as
part of e that mentions r.  Then the normal region subtyping kicks in.
For example, you can do "shallow" aliasing:

  ulist_t x = NULL;
  alias <`r> lh = (_ *`r)x in {
    lh = new List(1,NULL); // OK
    ...
  }

Here ulist_t is a list of unique pointers.  In this case, only the
outermost pointer is made non-linear, while the remainder of the
list is still unique.  You can also do "deep" aliasing if you cast
to a const structure:

  list_t<int,`U> u = NULL;
  alias <`r> x = (clist_t<int,`r>)u in {
    x = new CList(1,NULL);
    ...
  }

Here, list_t is a normal list, and clist_t is a constant list, where
the tail pointer is const.

Changing the syntax to be more C-like should be simple (not sure
if Greg or Trevor were going to tackle this ...): just do

  alias <`r> x = e;

In a lot of ways, this resembles the way we open existentials; I think
it seems like a nice mechanism.

One other change:
Started defining the idea of a "unique place" in the flow analysis.
This is not being used yet, but I'll finish it up when I find more time.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2325 581db07f-6e62-4d40-9733-d432c4eebb07