As per Trevor's suggestion, one now writes "@extensible datatype" instead of "datatype @extensible" or "datatype Foo @extensible". Similarly, one writes "@tagged union Foo" instead of "union Foo @tagged".
I also made changes that allow you to not write @extensible on every
occurrence of a datatype.  In particular, as long as the first
definition is declared as @extensible, then the subsequent definitions
(of the same datatype) are considered @extensible.  Similarly, a free
reference to an @extensible definition need not mention the
@extensible.  In other words, you can write "datatype exn" instead of
"@extensible datatype exn".

Modified the compiler so that internally, we use Datatype instead of
Tunion (e.g., Datatype_e instead of Tunion_e.)

And Dan was right -- he did tell me about the problems with flat
tunions earlier and I actually had the code in there to prevent the
problem.  Urgh.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2333 581db07f-6e62-4d40-9733-d432c4eebb07