Bug fixes to help Mike out with Medianet.
* Fixed bug in absyndump for printing the unique region

* Fixed bug in sub-typing code that was preventing casting a
T** to a T*const*const.

* Fixed bug in toc where we were taking the address of translated
expressions that might have casts in them.  We now push the
& inside the casts.  (Fixed a bug with the swap operation).

* Maybe fixed a bug in rgnorder:  If we had a function:

int foo(int *`r::TR) { ... }

then `r was not treated as outliving `foo because it's
of kind TR.  I don't see why this restriction is needed
so I added the `r outlives `foo constraint.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2347 581db07f-6e62-4d40-9733-d432c4eebb07