Cleaned up the xml mess.
For those that want to know: the procedure for changing a
library header file is to copy the old version from its
canonical location, include/, into lib/, then make changes
there.  This allows booting (make) to use the old version
and make cyclone_src use the new version.  On make update
the new version is moved onto the old version.  This means
.h files should not appear in lib/ in the CVS repository.
But the xml files WERE in the CVS repository in lib/,
probably for historical reasons.  So the old version kept
getting restored on every cvs update.  The solution: get
rid of them from the repository.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2360 581db07f-6e62-4d40-9733-d432c4eebb07