Changed kinds to be a product of two qualifiers: the old kind part and an additional alias qualifier, which is one of Aliasable, Unique, or Top.  The old kinds are now essentially k,Aliasable for all k except when k == RgnKind, for which we have RgnKind,Top == TopRgnKind, RgnKind,Unique == UniqueRgnKind, RgnKind,Aliasable == RgnKind.  Created a lot of constants to avoid excessive allocation; see tcutil.h.  Concrete syntax is as before, but with U or T preceding the value.  E.g. TB indicates BoxKind,Top.
I have yet to take advantage of this flexibility.  The next step
is to extend this product to BoxKind, MemKind, and AnyKind, so
that polymorphic functions work correctly.  E.g. if we have
foo(`a::TB), then this means we can instantiate `a with any kind
of boxed type, unique or otherwise.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2451 581db07f-6e62-4d40-9733-d432c4eebb07