Made changes to allow parametric polymorphism on unique pointers.  The basic changes were simple:
- Change typ_kind to return either M or UM for ? pointers and B or UB for
  */@ pointers, depending on their region.
- Allow more general contexts for checking type well-formedness; e.g.
  pointer types have expected kind <= TM (rather than just M).

The challenge was in getting the inference to work well.  The approach I
ended up with is as follows.  Free universal type variables are given the
most general kind in type declarations (i.e. structs, datatypes, etc.) and
typedefs.  For example, inference will assign TR kind to region variables in
typedefs, or TB kind to value variables.  For example:

  struct List<`a,`r>{`a hd; struct List<`a,`r> *`r tl;};

Here `a will be given kind TB, and `r will be given kind TR.  Existential
variables default to having Aliasable kind.

In function parameters, type variables are given Aliasable kind by default.
For example:

  extern list_t<`b> map(`b f(`a),list_t<`a> x);

Here `b and `a are given kind B and the inferred region variable of the list
is given kind R.  You can add the kinds explicitly to specialize the
definitions to use unique pointers for more generality, e.g.:

  extern int length(list_t<`a::TB> x);

The utility of the top kinds (i.e. TB) is evident here, as the same length
function works for a list containing uniquely-pointed-to elements and a list
with aliasable elements.  You would not be able to do this with the alias
construct, since `a is nested and not const.

I had to be careful about not overly generalizing and pinning the types of
pointers, type variables, and evars.  For example, a type definition (like
List above) might generally allow a top-kinded type, e.g. TR for the list
pointer.  However, the context in which that type definition is used may
expect a more refined kind, e.g. struct List *`r could be expected to have
kind B rather than kind TB; this condition can be satisfied as long as `r
has kind TR rather than R, so I check for exactly these sorts of
situations.

I changed some of the List, Queue, and Core modules to be a bit more general
and to use these features.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2452 581db07f-6e62-4d40-9733-d432c4eebb07