Fixed a bug in is_noalias_path.  The problem is that it was considering pointer dereferences through pointers of Top RgnKind as non-aliasable.  This is not sound, as the following example shows:
  typedef int @`r @`r2 handle_t<`r,`r2>;

  void myfree(handle_t<`U,`r::TR> h) {
    let b = *h;
    Core::ufree(b);
  }

  void foo() {
    handle_t<`U,`H> h = new (new 1);
    myfree(h);
    myfree(h); // oops!
  }

Really what we want is for the type to be non-aliasable, but for all of the
pointers that are dereferenced to be *unique.* This was a quick fix.

Added code for "alias inference."  That is, we do a modest job of figuring
out where we might insert an alias construct so as not to consume a unique
pointer and then do so automatically.  We determine where an alias might be
inserted (as implemented by Tcutil::can_insert_alias) as follows:

   If an expression appears in a function argument that is a no-alias
   pointer and has a no-alias path, and the expected type is non-unique
   (i.e. it is either Top or Aliasable) but is not the Heap (i.e. it is a
   region variable), then we can insert a coercion.  Because this test
   relies on the types of a function's formal and actual arguments, we have
   to do it during typechecking (which complicates the next part).

To actually insert the alias coercion(s) (as implemented by
Tcexp::insert_alias_stmts), we basically want to rewrite the syntax to add
an alias statement surrounding the given function call as a statement
expression.  That is, if we have

y = f(e);

where e is a unique pointer, we would rewrite to (simplified)

y = ({ alias <`r> x = e; f(x); })

While this basic idea is simple, it's a pain to do in practice because
typechecking has lots of side-effects.  Since we have to typecheck to
determine where to insert alias statements, and we would like to typecheck
our newly introduced syntax (so as to properly prep it for flow analysis
etc.), this presents a problem.  The ad hoc solution currently in use is to
"deep-copy" all function call expressions before typechecking them, and if
in the process an opportunity is discovered to introduce alias statements,
the copied syntax is modified, typechecked, and inserted in the program
AST.  This works most of the time, but I've found cases where the copy is
apparently not good enough; I'll try to track down in time.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2472 581db07f-6e62-4d40-9733-d432c4eebb07