Another fix for uniqueness and tagged unions: to determine if a tagged union has been consumed, we refine the check to see if the last field written to is known; if so, we only look at that field. I also made some small fixes to join_absrval and after_absrval for Aggregates with regard to the fieldnum.  Now the following code works:
  @tagged union foo { int *`U w; int *`U z; };
  void fun5(union foo f) {
    Core::ufree(f.x);
    f.y = new 1;
    union foo g = f; // flow thinks that f.x consumed, but knows field is f.y
  }

However, the following code does not work

  void fun(union foo f, int b) {
    if (b) {
      Core::ufree(f.x);
      f.y = new 1;
    }
    else {
      Core::ufree(f.y);
      f.x = new 1;
    }
    union foo g = f;
  }

This is because on the join, f is considered to be completely consumed and
its field is unknown, even though at the end of each branch, we know that
whatever tag f has, its value is unconsumed.  I don't think this problem
can be fixed without moving away from using Aggregates to model tagged
unions, but I'm not sure the change is worth it.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2498 581db07f-6e62-4d40-9733-d432c4eebb07