Refined the flow analysis for tagged unions to be able to properly handle initializing different variants in different branches.  In particular, the following used to fail:
  @tagged union foo {
    int * x;
    int * y;
  };

  void fun(int b) {
    union foo f;
    if (b)
      f.x = new 1;
    else
      f.y = new 1;
  }

The join at the end of the if would have the current tag being unknown
and both fields of the variant being uninitialized!  To fix this,
whenever we assign to a field in a variant, we change the other
fields to unknown_all (propogating down through Aggregates).  The
result is that the above join still has unknown tag, but every field
is considered initialized.  This also works out well with unique
pointers.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2499 581db07f-6e62-4d40-9733-d432c4eebb07