I've fixed the bug Yanling reported.  This sort of thing has happened before, which is the only reason I knew what to look for.  Here goes:
* toc.cyc isn't very careful about sharing expressions because it
  wasn't written with a subsequent pass in mind.  
* tovc.cyc mutates expressions in place.

That's a recipe for disaster, of course, but nobody has had time to either:
* carefully remove all implicit sharing in code generated by toc.cyc
* have tovc.cyc be less destructive

Now, this only comes up when tovc creates a function, which in
practice only happens with comprehensions.  So the code in toc.cyc for
comprehensions was written carefully with a big warning that changing
it would mess up tovc.  But with the comprehension nested inside a
struct initializer, the struct initializer suffered from the same
problem.

Anyway, there are probably more bugs lurking out there.  What will
trigger them is a comprehension forcing some outer code to be in a
generated function.  I don't have time right now to come up with
something sane, but at least we all now know what the issue is.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2514 581db07f-6e62-4d40-9733-d432c4eebb07