* A few fixes and changes that let you declare top-level regions using an abstract typedef.  For instance:
typedef _::R myregion;

void foo(int *@region(myregion) x) { ... }

Note that you have to use the @region(...) syntax for such regions,
since int*myregion won't parse properly.  Given this, we could define
the unique region (`UC) and the reference counted region (`RC) in
Core:

typedef _::UR uniqueregion;
#define URGN @region(uniqueregion);
extern region_t<uniqueregion> unique_region;

This has not been tested extensively, so there may be bugs or gotchas
lurking.  Furthermore, there's no way to specify things such as
outlives constraints, or top-level (dynamic) region handles for them.
Still, this could be useful.

* Changed pretty printer (and absyndump) so they don't choke when a
non-variable, non-heap, non-unique, etc. type shows up as the region
for a given pointer.  While I was at it, changed the pretty printer to
use the extended @region(...) syntax except for regions where it is
unnecessary.

* Fixed bug in absyndump -- we have to print out typedefs due to
abstract typedefs.

* Fixed bug in unify_it for abstract typedefs -- was not allowing them
to unify.

* Fixed bug in tests/Makefile -- forgot to add flatlist to the list of
things to compile for "make all".  Also forgot to add flatlist to
"make clean".

* Fixed the code generator so that it can deal with the situation
where a struct whose last field is an abstract type variable that is
existentially quantified can be created.  For instance:

  struct Foo { <`a::A> int x; `a y; };

  struct Foo@ string_foo() {
    return new Foo{<const char ?> .x = 0, .y = "hello"};
  }

This is useful for simulating objects without a level of indirection
for the "private" data.  In this case, the translation must generate a
new struct so we get something like this as output:

  struct Foo { int x; };  // abstract last field is dropped

  struct Foo *string_foo() {
    // we use an anonymous struct here for the actual type
    struct {int x; struct _dyneither_ptr y;} *temp;
    temp = malloc(sizeof(*temp));
    temp->x = 0;
    temp->y = ...
    // after creating and initializing it, we cast up to Foo
    return (struct Foo*)temp;
  }

* Got rid of warning in rope.cyc via explicit cast.

* Got rid of two warnings in cycstubs.cyc via an explicit cast in one case,
and rewriting the code in another.  While I was at it, region-allocated
argument to execlp.

* Got rid of warning in cyclex/lexer.cyl and bison/reader.cyc.

* Got rid of warnings in xmlscan.cyl.

* Gor rid of warning in scanf.cyc.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2584 581db07f-6e62-4d40-9733-d432c4eebb07