Warning:  I'd suggest checking this version out in a fresh directory and making sure it builds.  There were some pretty substantial changes to cyc_include.h which may have introduced a bug for non-Cygwin, though I've tried to test things to rule this out.
We now track (in lex.cyl) a mapping from positions in the .cyp file to
source file and line numbers.  The mapping is sparse so it doesn't use
up much space, and it's very fast relative to using the lineno library
(which re-opens and lexes the files to do the mapping.)  (It also uses
unique pointers to support resizeable arrays.)

This information is used to emit .cyc source location information in
the absyndump routines (and replaces James's older code based on a
batch lineno call.)  As a result, we have fairly accurate source
location information in the generated .c files.  I say fairly accurate
because it's not possible to emit # directives within macros, so
I'm basically being conservative and only emitting them around
statements, declarations, and a few expression forms, making sure
that if a "function call" in the abstract syntax looks like a macro,
then we don't dump out any location info while dumping the arguments.
Bottom line:  the location info is accurate to the statement level.

This enables a number of things:

* GDB now uses the .cyc file for source, and allows you to step
through it.  It's still somewhat confusing to do this because of
namespace mangling and the re-mapping of local variables to temps done
in the translation.  Still, it works surprisingly well.

* I presume that the allocation profiler and gprof give reports in terms
of the original source files instead of the C files, since they
used __FILE__ and __LINE__.

* When we throw an exception, we're able to record the file and line
number so that we can print them out when an exception isn't caught.
(This was the real purpose behind all of this...)  Now, for instance,
you get a message like:

  Uncaught exception Cyc_Null_Exception thrown from around foo.cyc:16

To do the exception thing well, I had to change all of the exception
throwing functions (e.g., _throw, _throw_null, etc.)  into macros
which call functions, passing in the __FILE__ and __LINE__.  In
addition, I had to modify all of the other functions in cyc_include.h
which wanted to throw an exception so that they took the __FILE__ and
__LINE__ as extra arguments and then wrap the functions in a macro to
insert this info at the call site.  This was necessary so that, for
instance, on the Mac (which uses static inlined functions instead of
macros for most fat or zero-terminated pointer manipulation) we'd get
the location information for the call site instead of the
pointer-manipulating function in cyc_include.

Note that the pretty printer does not yet dump out location
information.  You only get this stuff when you use absyndump.  Note
also that I've eliminated the --lineno flag and replaced it with a
--nolineno flag, since the default behavior is to emit the location
info in the .c file.

This does slow down compilation a bit, but not by much and it
probably results in somewhat larger code sizes.  But personally,
I like having the information there by default.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2619 581db07f-6e62-4d40-9733-d432c4eebb07