Added _profile_region_calloc() to the runtime and a macro in cyc_include.h to call this when CYC_REGION_PROFILE is turned on.
Modified aprof so it doesn't die when it finds that something is freed
which was not recorded.  This seems to happen for all heap objects.
Also, modified aprof to give a linenumber in the amon.out file when an
error occurs.

We weren't treeshaking an extern declaration if it occurred twice
(since we successfully looked it up in the environment, it was
recorded as "used".)  This should now be fixed.

We now have limited support for qualified declarations.  In
particular, you can use Cyc:: or C:: to specify an absolute namespace
for a declaration.  For instance, if you have:

namespace Foo {
  extern int Cyc::i;
}

then i is not defined in namespace Foo.  Rather, it's defined at
top-level in the Cyclone namespace (as opposed to the C namespace.)
Note, however, that the lexer will barf if you try to declare a
variable in a namespace it's never seen.  For instance, if you wrote:

extern int Cyc::Bar::i;

then it's not going to be happy, though if you wrote:

namespace Bar { }
extern int Cyc::Bar::i;

it would be.  Obviously, I need to make this a bit better.  I did
what was minimal in order to support re-reading back in an interface
file (see below).

Fixed lots of bugs in absyndump for spitting out Cyclone code.
For instance, it wasn't properly printing out effects or kinds.
In addition, it now rewrites internal type variables (e.g., #14)
to something that can be read back in (e.g., `TTT14).  This needs
to be fixed so that we avoid renaming clashes.

Added a routine to absyndump.cyc to spit out an "interface" to and
added a --geninterface flag.  The interface consists of all of the
top-level declarations in <file> minus initializing expressions for
variables and function bodies, and also minus any extern variables
that are shaken out after type-checking.  These declarations are
dumped into <file>.cyci and we should now be able to re-parse these.
(A few other declarations are dropped --- typedefs are expanded out
[since we can't have a typedef defined twice with the same name] and
extern "C include" declarations aren't properly handled yet.)

The long term goal is to use this in lieu of the interface.cyc code
which has bit-rotted to the point where it's no longer usable.  In
particular, we should be able to take foo.cyc, generate foo.cyci and
then type-check their concatenation.  The next step is to provide
support at link time for sucking in all of the .cyci files and
to type-check them for consistency.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2627 581db07f-6e62-4d40-9733-d432c4eebb07